backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt>=4.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.5.0
alembic==1.13.0
passlib[bcrypt]
python-multipart
python-slugify==8.0.4
PyJWT==2.8.0
python-dateutil==2.8.2
